Updaing messaging and ensuring proper token

+++ handler/oauth2/migration.go
@@ -24,17 +24,24 @@ func (c *TokenMigrationHandler) MigrateToken(ctx context.Context, requester fosi
 	if token := responder.GetExtra("refresh_token"); token != nil {
 		var ok bool
 		if refreshToken, ok = token.(string); !ok {
-			return fosite.ErrInvalidRequest
+			return errors.Wrap(fosite.ErrInvalidRequest, "The refresh token is in a bad format")
 		}
 	}
 
 	tokenSignature := c.AccessTokenStrategy.AccessTokenSignature(responder.GetAccessToken())
+
+	if tokenSignature == "" {
+		return errors.Wrap(fosite.ErrInvalidRequest, "The access token is in a bad format")
+	}
 
 	if err = c.AccessTokenStorage.CreateAccessTokenSession(ctx, tokenSignature, requester); err != nil {
 		return err
 	}
 	if refreshToken != "" {
 		refreshTokenSignature := c.RefreshTokenStrategy.RefreshTokenSignature(refreshToken)
+		if refreshTokenSignature == "" {
+			return errors.Wrap(fosite.ErrInvalidRequest, "The refresh token is in a bad format")
+		}
 		if err = c.RefreshTokenStorage.CreateRefreshTokenSession(ctx, refreshTokenSignature, requester); err != nil {
 			if ex := c.AccessTokenStorage.DeleteAccessTokenSession(ctx, tokenSignature); ex != nil {
 				err = errors.Wrap(err, ex.Error())
